requirements.txt

os

python-dotenv
